refactor: 예외처리 클래스 추가 및 적용

+++ src/main/java/org/example/newsfeedproject/comment/service/CommentService.java
@@ -4,14 +4,18 @@ import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.example.newsfeedproject.comment.dto.CommentResponse;
 import org.example.newsfeedproject.comment.entity.Comment;
+import org.example.newsfeedproject.comment.exception.CommentNotFoundException;
+import org.example.newsfeedproject.comment.exception.CommentUnauthorizedException;
 import org.example.newsfeedproject.comment.repository.CommentRepository;
 import org.example.newsfeedproject.feed.entity.Feed;
+import org.example.newsfeedproject.like.repository.CommentLikeRepository;
 import org.example.newsfeedproject.user.entity.User;
 import org.example.newsfeedproject.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -19,6 +23,7 @@ import java.util.Optional;
 public class CommentService {
 
       private static CommentRepository commentRepository;
+      private static CommentLikeRepository commentLikeRepository;
 //    private static FeedRepository feedRepository;
 //    private static UserRepository userRepository;
 //
@@ -44,19 +49,24 @@ public class CommentService {
                 .comments(comments)
                 .build();
         Comment saved = commentRepository.save(comment);
-        return Comment.toDto(saved);
+
+        // 처음 작성 시 초기값 등록
+        return CommentResponse.of(saved, false, 0L);
     }
 
     @Transactional
     public List<CommentResponse> getAllComments(Long feedId, Long userId){
-        final List<CommentResponse> commentList = commentRepository.findAllByFeedId(feedId, userId);
-        return commentList;
+        // 피드 존재 여부 확인
+//        if (!feedRepository.existsById(feedId)) {
+//            throw new FeedNotFoundException(feedId);
+//        }
+        return commentRepository.findAllByFeedId(feedId, userId);
     }
 
     @Transactional
     public CommentResponse getComment(Long id){
         final Optional<Comment> optionalComment = commentRepository.findById(id);
-        if(optionalComment.isEmpty()) throw new RuntimeException("존재하지 않은 댓글입니다.");
+        if(optionalComment.isEmpty()) throw new CommentNotFoundException(id);
         return Comment.toDto(optionalComment.get());
     }
 
@@ -65,8 +75,8 @@ public class CommentService {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow();
 
-        if(comment.getUser().getId() != userId)
-            throw new RuntimeException("자신의 댓글만 수정할 수 있습니다.");
+        if(!Objects.equals(comment.getUser().getId(), userId))
+            throw new CommentUnauthorizedException();
         comment.setComments(comments);
 
         commentRepository.save(comment);
@@ -75,7 +85,19 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteComment(Long id){
+    public void deleteComment(Long id,Long userId){
+        Comment comment = commentRepository.findById(id)
+                .orElseThrow();
+
+        if(!Objects.equals(comment.getUser().getId(), userId))
+            throw new CommentUnauthorizedException();
+
         commentRepository.deleteById(id);
+    }
+
+    // 피드 삭제 시 동시 삭제
+    @Transactional
+    public void deleteAllComments(Long feedId){
+        commentRepository.deleteByFeedId(feedId);
     }
 }
